Adding MMSegmentation and Refactor on relative path for weigh
