--- adv_manhole/models/depth_hints/networks/__init__.py
@@ -1,4 +1,0 @@
-from .resnet_encoder import ResnetEncoder
-from .depth_decoder import DepthDecoder
-from .pose_decoder import PoseDecoder
-from .pose_cnn import PoseCNN
