--- adv_manhole/models/many_depth/networks/__init__.py
@@ -1,4 +1,0 @@
-from .resnet_encoder import ResnetEncoder, ResnetEncoderMatching
-from .depth_decoder import DepthDecoder
-from .pose_decoder import PoseDecoder
-from .pose_cnn import PoseCNN
