+++ adv_manhole/models/mm_segmentation/.gitignore
@@ -1,0 +1,2 @@
+weights/*
+vis_data/*
