+++ adv_manhole/models/monodepth2/.gitignore
@@ -1,0 +1,1 @@
+weights/*
