+++ adv_manhole/models/monodepth2/intrinsics/test_sequence_intrinsics.json
@@ -1,0 +1,5 @@
+[
+    [0.58, 0, 0.5],
+    [0, 1.92, 0.5],
+    [0, 0, 1.0]
+]
