+++ requirements.txt
@@ -5,3 +5,7 @@ torch
 torchvision
 super-gradients
 wandb
+openmim
+mmengine
+mmcv==2.1.0
+mmsegmentation
